Rename random.* to util.*
