+++ test/unittests/active_info.c
@@ -1,7 +1,7 @@
 // Copyright 2016 ELIFE. All rights reserved.
 // Use of this source code is governed by a MIT
 // license that can be found in the LICENSE file.
-#include "random.h"
+#include "util.h"
 #include <inform/active_info.h>
 #include <math.h>
 #include <unit.h>
