+++ test/unittests/util.c
@@ -1,7 +1,7 @@
 // Copyright 2016 ELIFE. All rights reserved.
 // Use of this source code is governed by a MIT
 // license that can be found in the LICENSE file.
-#include "random.h"
+#include "util.h"
 
 int* random_series(size_t size, int base)
 {
